refact: move room store to stores folder
