+++ src/components/domain/connection-dialog/index.tsx
@@ -2,8 +2,7 @@ import classNames from "classnames";
 import { useRouter } from "next/router";
 import { useEffect, useRef, useState } from "react";
 import toast from "react-hot-toast";
-import { useRoomStore } from "../../../contexts/room-context/room-store";
-import { MainRoomEvents } from "../../../contexts/room-context/types";
+import { useRoomStore, MainRoomEvents } from "../../../stores/room-store";
 import Dialog from "../../ui/dialog";
 import { ConnectingMessage } from "./connecting-message";
 import { PeopleForm } from "./people-form";
